more refac
