+++ src/components/SearchBox/utils.tsx
@@ -2,6 +2,8 @@ import styled from 'styled-components';
 import { Grid, Typography } from '@mui/material';
 import React from 'react';
 import maplibregl from 'maplibre-gl';
+import match from 'autosuggest-highlight/match';
+import parse from 'autosuggest-highlight/parse';
 import { useMapStateContext } from '../utils/MapStateContext';
 import { t } from '../../services/intl';
 import { getGlobalMap } from '../../services/mapStorage';
@@ -67,4 +69,18 @@ export const fitBounds = (option, panelShown = false) => {
       padding: { top: 5, bottom: 5, right: 5, left: panelWidth + 5 },
     });
   }
+};
+
+export const highlightText = (resultText, inputValue) => {
+  const matches = match(resultText, inputValue, {
+    insideWords: true,
+    findAllOccurrences: true,
+  });
+  const parts = parse(resultText, matches);
+  return parts.map((part, index) => (
+    // eslint-disable-next-line react/no-array-index-key
+    <span key={index} style={{ fontWeight: part.highlight ? 700 : 400 }}>
+      {part.text}
+    </span>
+  ));
 };
